requirements.txt
selenium==4.15.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
pyarrow==14.0.1
matplotlib==3.8.2
seaborn==0.13.0
nltk==3.8.1
textblob==0.17.1
python-dotenv==1.0.0
pydantic==2.5.0
loguru==0.7.2
tqdm==4.66.1
aiohttp==3.9.1
aiofiles==23.2.1
webdriver-manager==4.0.1
pytz==2023.3
cachetools==5.3.2
torch==2.1.2
transformers==4.36.2
pytorch-lightning==2.1.3